Phase 4.4: extract storage user route groups
